Декомпозирует MainScreen
